+++ benchmarks/Transformers/L2NormalizerBench.php
@@ -40,7 +40,7 @@ class L2NormalizerBench
     /**
      * @Subject
      * @Iterations(3)
-     * @OutputTimeUnit("seconds", precision=3)
+     * @OutputTimeUnit("milliseconds", precision=3)
      */
     public function apply() : void
     {