+++ src/Transformers/GaussianRandomProjector.php
@@ -127,9 +127,7 @@ class GaussianRandomProjector implements Transformer, Stateful, Stringable
             throw new RuntimeException('Transformer has not been fitted.');
         }
 
-        $samples = Matrix::quick($samples)
-            ->matmul($this->r)
-            ->asArray();
+        $samples = Matrix::quick($samples)->matmul($this->r)->asArray();
     }
 
     /**