+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - 0.2.0
+    - Added Flysystem persister for S3, Azure, FTP, etc.
     - Add Recursive Feature Eliminator feature selector
     - Can now disable holdout validation in MLP learners
     - Added instability detection to gradient-based learners